requirements.txt
click==7.1.2
databricks-connect==6.5.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
PyYAML==5.1.1
